backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.6
pydantic-settings>=2.1
passlib[bcrypt]>=1.7
bcrypt>=4.0
email-validator>=2.1
python-jose[cryptography]>=3.3
python-dotenv>=1.0
alembic>=1.13

# Testing dependencies
pytest>=7.4
pytest-cov>=4.1
pytest-mock>=3.12
pytest-asyncio>=0.21
httpx>=0.25
factory-boy>=3.3
faker>=20.1